feat(tokens): Step 6a+6b — widen write path for u128 lib-client
lib-client transaction builders (build_sov_wallet_transfer_tx,
build_token_wallet_transfer_tx, build_mint_tx, build_create_token_tx,
build_burn_tx) now take u128 amount parameters. This commit updates the
mobile-side Rust glue and the remaining TS write-path call sites:

- quic-jni/src/identity_bridge.rs: 4 call sites `amount → amount as u128`
- quic-jni/src/identity_jni.rs: 6 call sites `amount as u64 → amount as u64 as u128`
  (jlong is signed i64; bit-preserve to u64 first, then widen to u128)
- Cargo.lock: updated after lib-client dependency rebuild

JNI surface (jlong) and Kotlin Long types are unchanged — the mobile write
path still caps at i64::MAX atoms ≈ 9.22 SOV at 18 decimals, but fails
cleanly at the boundary rather than silently corrupting. Real widening to
u128-capable string params is deferred to a follow-up.

- src/services/TokenService.ts: createToken passes String(initial_supply)
  and String(max_supply) through the bridge (bridge types already expected
  strings; Number() cast was a leftover)
- src/screens/SendTokensScreen.tsx: balance parsing now uses atomsToNumber
  (bigint-safe); tokens missing decimals are skipped with a console warn
  instead of silently defaulting to 8; SOV decimals hardcoded at 18 where
  the endpoint is SOV-only
- src/utils/tokenUnits.ts: add atomsToNumber(atoms, decimals) for legacy
  number-typed call sites that need a precision-safe bigint→number path

Native libs rebuilt externally via scripts/build-lib-client-ios.sh and
quic-jni/build-android.sh (all three Android ABIs + iOS arm64). Artifacts
are gitignored.

+++ android/app/src/main/rust/quic-jni/src/identity_bridge.rs
@@ -181,7 +181,7 @@ pub fn build_token_create_transaction(
         &identity,
         name,
         symbol,
-        initial_supply,
+        initial_supply as u128,
         decimals,
         treasury_recipient,
         chain_id,
@@ -213,7 +213,7 @@ pub fn build_token_mint_transaction(
         to_arr[..to_pubkey.len()].copy_from_slice(to_pubkey);
     }
 
-    zhtp_client::build_mint_tx(&identity, &token_id_arr, &to_arr, amount, chain_id)
+    zhtp_client::build_mint_tx(&identity, &token_id_arr, &to_arr, amount as u128, chain_id)
         .map_err(|e| anyhow::anyhow!("Failed to build token mint transaction: {}", e))
 }
 
@@ -242,7 +242,7 @@ pub fn build_token_transfer_transaction(
         to_arr[..to_pubkey.len()].copy_from_slice(to_pubkey);
     }
 
-    zhtp_client::build_transfer_tx(&identity, &token_id_arr, &to_arr, amount, chain_id, nonce)
+    zhtp_client::build_transfer_tx(&identity, &token_id_arr, &to_arr, amount as u128, chain_id, nonce)
         .map_err(|e| anyhow::anyhow!("Failed to build token transfer transaction: {}", e))
 }
 
@@ -262,6 +262,6 @@ pub fn build_token_burn_transaction(
         token_id_arr[..token_id.len()].copy_from_slice(token_id);
     }
 
-    zhtp_client::build_burn_tx(&identity, &token_id_arr, amount, chain_id)
+    zhtp_client::build_burn_tx(&identity, &token_id_arr, amount as u128, chain_id)
         .map_err(|e| anyhow::anyhow!("Failed to build token burn transaction: {}", e))
 }
